python module poc
