+++ e2e/setup/basic-auth-executor-config.yaml
@@ -1,0 +1,5 @@
+apiConnection:
+  armadaUrl: "server:50051"
+  basicAuth:
+    Username: test
+    Password: asdfasdf
