+++ e2e/setup/secure-basic-armada-auth-config.yaml
@@ -1,0 +1,18 @@
+auth:
+  basicAuth:
+    enableAuthentication: true
+    Username: test
+    Password: asdfasdf
+  anonymousAuth: true
+  permissionGroupMapping:
+    submit_jobs: ["everyone"]
+    submit_any_jobs: ["everyone"]
+    create_queue: ["everyone"]
+    delete_queue: ["everyone"]
+    cancel_jobs: ["everyone"]
+    cancel_any_jobs: ["everyone"]
+    reprioritize_jobs: ["everyone"]
+    reprioritize_any_jobs: ["everyone"]
+    watch_events: ["everyone"]
+    watch_all_events: ["everyone"]
+    execute_jobs: ["everyone"]
